128-bit constants; the libraries on wide types; binary128 from the same bodies
Inst::IConst carries an i128 and a `const` width expression evaluates at
128 bits, so `const 1 << M` is right for M = 112 (it used to wrap at 64
and put the hidden bit at 48). A 64-bit pack literal keeps accepting
any 64-bit pattern.

lib/float.ssa: mul forms the exact 2M + 2-bit product in u(2M + 10);
fma's accumulator is u(2M + 20); float-to-float conv re-expresses the
significand in u(M + N + 10); fint and fmag place a u64 magnitude in
u(M + 69). For f32 those are still one word; for f64 two; and one body
now serves float(15, 112). lib/fixed.ssa and lib/unit.ssa multiply and
divide in u128. The hand-rolled add128/sub128/shl128/shr128/msb128/
lt128/udiv128 helpers are gone; lib/wide.ssa keeps only the division
generics the lowering dispatches to.

suite/f128.ssa: 41 binary128 cases — add, sub, mul, div, sqrt, fma,
conv to and from f64 — with expectations computed exactly with Python
fractions and rounded to nearest even at 113 bits. 645/645 on native,
wasm, riscv, arm-qemu and --soft; TestFloat still 0 wrong on the
rewritten ops; the fuzzer agrees.

+++ src/emit.rs
@@ -895,7 +895,7 @@ fn compile_inst(e: &mut FnEmit, inst: &Inst) -> Result<(), String> {
             let r = e.repr(*dst);
             // materialize the canonical form; a 32-bit container only
             // needs its low half
-            let mut v = crate::opt::norm(r, *imm) as u64;
+            let mut v = crate::opt::norm(r, *imm as i64) as u64;
             if r.container() == 32 {
                 v &= 0xffff_ffff;
             }
